Added bootbox and modal confirming deleting a contact

+++ app/js/controllers/ContactDetailCtrl.js
@@ -3,13 +3,17 @@
 addressBookControllers.controller('ContactDetailCtrl', ['$scope', '$http','$routeParams','$location',
   function($scope, $http, $routeParams, $location){
     $http.get('http://fast-gorge.herokuapp.com/contacts/' + $routeParams.id).success(function(data){
-      $scope.contact = data
+      $scope.contact = data;
     })
 
-   $scope.deleteContact = function(contactID) {
-     $http.delete('http://fast-gorge.herokuapp.com/contacts/' + contactID).success(function(data){
-       $location.path("/contacts");
-     })
+    $scope.deleteContact = function(contactID) {
+      bootbox.confirm("Are you sure you want to delete this contact?", function(answer) {
+        if (answer === true) {
+          $http.delete('http://fast-gorge.herokuapp.com/contacts/' + contactID).success(function(data){
+            $location.path("/contacts");
+          })
+        }
+      })
    }
   }
 ]);
